fixed dependency bugs, and get all avatars now works
